+++ inc/template-tags.php
@@ -502,20 +502,6 @@ if ( ! function_exists( 'onepress_get_section_about_data' ) ) {
             }
         }
 
-        if (empty($page_ids)) {
-            $current_pos_id = get_the_ID();
-            $args = array(
-                'posts_per_page' => 3,
-                'orderby' => 'date',
-                'order' => 'DESC',
-                'exclude' => $current_pos_id,
-                'post_type' => 'page',
-            );
-            $posts_array = get_posts($args);
-            foreach ($posts_array as $p) {
-                $page_ids[] = array('content_page' => $p->ID, 'enable_link' => 0, 'hide_title' => 0);
-            }
-        }
         return $page_ids;
     }
 }
@@ -556,26 +542,17 @@ if ( ! function_exists( 'onepress_get_section_services_data' ) ) {
                 if (isset ($v['content_page'])) {
                     $v['content_page'] = absint($v['content_page']);
                     if ($v['content_page'] > 0) {
-                        $page_ids[] = wp_parse_args($v, array('icon' => 'gg', 'enable_link' => 0));
+                        $page_ids[] = wp_parse_args($v, array(
+                            'icon_type' => 'icon',
+                            'image' => '',
+                            'icon' => 'gg',
+                            'enable_link' => 0
+                        ));
                     }
                 }
             }
         }
         // if still empty data then get some page for demo
-        if (empty($page_ids)) {
-            $current_pos_id = get_the_ID();
-            $args = array(
-                'posts_per_page' => 4,
-                'orderby' => 'date',
-                'order' => 'DESC',
-                'exclude' => $current_pos_id,
-                'post_type' => 'page',
-            );
-            $posts_array = get_posts($args);
-            foreach ($posts_array as $p) {
-                $page_ids[] = array('content_page' => $p->ID, 'icon' => 'gg', 'enable_link' => 0);
-            }
-        }
         return $page_ids;
     }
 }