+++ section-parts/section-services.php
@@ -49,9 +49,22 @@ if ( ! empty( $page_ids ) ) {
                             $post = get_post($post_id);
                             setup_postdata($post);
                             $settings['icon'] = trim($settings['icon']);
-                            if ($settings['icon'] != '' && strpos($settings['icon'], 'fa-') !== 0) {
-                                $settings['icon'] = 'fa-' . $settings['icon'];
+
+                            $media = '';
+
+                            if ( $settings['icon_type'] == 'image' && $settings['image'] ){
+                                $url = onepress_get_media_url( $settings['image'] );
+                                if ( $url ) {
+                                    $media = '<div class="service-image icon-image"><img src="'.esc_url( $url ).'" alt=""></div>';
+                                }
+                            } else if ( $settings['icon'] ) {
+                                $settings['icon'] = trim( $settings['icon'] );
+                                if ($settings['icon'] != '' && strpos($settings['icon'], 'fa-') !== 0) {
+                                    $settings['icon'] = 'fa-' . $settings['icon'];
+                                }
+                                $media = '<div class="service-image"><i class="fa '.esc_attr( $settings['icon'] ).' fa-5x"></i></div>';
                             }
+
                             $classes = 'col-sm-12 col-md-6 col-lg-'.$layout;
                             if ($j >= $columns) {
                                 $j = 1;
@@ -77,11 +90,9 @@ if ( ! empty( $page_ids ) ) {
                                             ?>
                                         </div>
                                     <?php } ?>
-                                    <?php if ( $settings['icon'] != '' ) { ?>
-                                        <div class="service-image">
-                                            <i class="fa <?php echo esc_attr($settings['icon']); ?> fa-5x"></i>
-                                        </div>
-                                    <?php } ?>
+                                    <?php if ( $media != '' ) {
+                                        echo $media;
+                                    } ?>
                                     <div class="service-content">
                                         <h4 class="service-title"><?php the_title(); ?></h4>
                                         <?php the_excerpt(); ?>