+++ single.php
@@ -22,10 +22,10 @@ get_header(); ?>
 					<?php get_template_part( 'template-parts/content', 'single' ); ?>
 
 					<?php
-						// If comments are open or we have at least one comment, load up the comment template.
-						if ( comments_open() || get_comments_number() ) :
-							comments_template();
-						endif;
+					// If comments are open or we have at least one comment, load up the comment template.
+					if ( comments_open() || get_comments_number() ) :
+						comments_template();
+					endif;
 					?>
 
 				<?php endwhile; // End of the loop. ?>