+++ changelog.md
@@ -1,6 +1,14 @@
 ====================================================================
 CHANGELOG
 ====================================================================
+## 1.1.9
+* Add Hero options settings
+* Remove default service data.
+* Remove default about data.
+* Fix feature items break layout.
+* Add Features/Services item image.
+
+
 ## 1.1.8
 * Update pot file.
 * Change text area to JS Editor control.